+++ Classes/common/CDTDatastore.m
@@ -313,7 +313,7 @@ NSString* const CDTDatastoreChangeNotification = @"CDTDatastoreChangeNotificatio
 
         // Deleted documents won't have a `doc` field
         if (!deleted) {
-        revision.body = [[TD_Body alloc] initWithProperties:row[@"doc"]];
+            revision.body = [[TD_Body alloc] initWithProperties:row[@"doc"]];
         }
 
         CDTDocumentRevision *ob = [[CDTDocumentRevision alloc] initWithTDRevision:revision];
@@ -387,9 +387,9 @@ NSString* const CDTDatastoreChangeNotification = @"CDTDatastoreChangeNotificatio
 }
 
 
--(BOOL) deleteDocumentWithId:(NSString*)docId
-                         rev:(NSString*)rev
-                       error:(NSError * __autoreleasing *)error
+-(CDTDocumentRevision*) deleteDocumentWithId:(NSString*)docId
+                                         rev:(NSString*)rev
+                                       error:(NSError * __autoreleasing *)error
 {
     if (![self ensureDatabaseOpen]) {
         *error = TDStatusToNSError(kTDStatusException, nil);
@@ -400,16 +400,16 @@ NSString* const CDTDatastoreChangeNotification = @"CDTDatastoreChangeNotificatio
                                                          revID:nil
                                                        deleted:YES];
     TDStatus status;
-    [self.database putRevision:revision
-                prevRevisionID:rev
-                 allowConflict:NO
-                        status:&status];
+    TD_Revision *new = [self.database putRevision:revision
+                                   prevRevisionID:rev
+                                    allowConflict:NO
+                                           status:&status];
     if (TDStatusIsError(status)) {
         *error = TDStatusToNSError(status, nil);
-        return NO;
+        return nil;
     }
 
-    return YES;
+    return [[CDTDocumentRevision alloc] initWithTDRevision:new];
 }
 
 -(NSString*) extensionDataFolder:(NSString*)extensionName